[PROJ] clinkedlist -> linked_list

+++ kernel/kernel_processes/boot/bootprocess_sm.cpp
@@ -16,7 +16,7 @@ void BootSM::initialize(){
     disable_visual();
     usb_start_polling();
     gpu_size screen_size = gpu_get_screen_size();
-    mouse_config((gpu_point){screen_size.width/2,screen_size.height/2}, screen_size);
+    mouse_config((gpu_point){(i32)screen_size.width/2,(i32)screen_size.height/2}, screen_size);
     AdvanceToState(Bootscreen);
 }
